+++ integrations/github/github/core/options.py
@@ -1,4 +1,4 @@
-from typing import TypedDict
+from typing import Any, Required, TypedDict
 
 
 class SingleRepositoryOptions(TypedDict):
@@ -11,3 +11,14 @@ class ListRepositoryOptions(TypedDict):
     """Options for listing repositories."""
 
     type: str
+
+
+class SingleFolderOptions(TypedDict):
+    repo: str
+    path: str
+
+
+class ListFolderOptions(TypedDict):
+    repo: Required[dict[str, Any]]
+    path: Required[str]
+    branch: Required[str]