+++ integrations/github/CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped ocean version to ^0.24.7
 
+### Features
+
+- Add support for Github folder kind
+
 
 ## 0.1.10-dev (2025-06-09)
 
@@ -81,7 +85,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.3-dev (2025-06-01)
 
-
 ### Improvements
 
 - Bumped ocean version to ^0.23.5
@@ -103,7 +106,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - Bumped ocean version to ^0.23.4
-
 
 ## 0.1.0-dev (2025-05-28)